Add more complete LaTeX stack to better support publishing
This adds key tools that myst regularly uses to build PDFs.  They are just regular TeX packages that are likely to be useful to users beyond the myst toolchain.

This is still a small subset of all possible texlive packages, we may get more requests for a good minimal TeX stack as people begin using it more.